Fix integration tests that should be unit tests
